Finishing handling max and min values for %d in combination with any flags

+++ libft/srcs/ft_abs.c
@@ -12,7 +12,7 @@
 
 #include "libft.h"
 
-int				ft_abs(long long i)
+int				ft_abs(intmax_t i)
 {
 	if (i < 0)
 		return (i * -1);
